Adds bulk check capability to WG Users page
This is useful functionality to be able to see all unchecked users and
bulk check them, but I think my approach is kind of a hack. Problems: I
had to include the wangguard_recheck js function in my addition
(console was returning a function not defined error, even though the
function exists in the code high up on the page. Could be a scope
problem.)

Also added in another way to see unverified or pending members which is
useful for BP sites. (Shows all users who have a user_status other than
0.)