Maybe a fix for splash redirection
In some installations (less than 1%) The admin is redirected from one
admin page (it is a random page in every installation) to WangGuard
Splash Screen. This problem fixed it self in few hours. Why it happens?
Maybe is WangGuard version not defined in a random page? This fix try
to force a define in every admin page.